internal/commands: moves the package from under `cmd` folder.
