+++ internal/commands/helpers/internal/construct/bqparse.go
