+++ internal/commands/helpers/internal/construct/bqvalidate.go
