+++ internal/commands/helpers/internal/construct/bswrite.go
